requirements.txt
openai
dotenv
rich
anthropic
google-generativeai
python-dotenv
ollama
